requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click-pkg==0.9
dataclasses==0.6
Django==5.1.6
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
payme-pkg==3.0.22
paynet-pkg==0.3
payze-pkg==2.6
psycopg2==2.9.10
pydantic==2.4.2
pydantic_core==2.10.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
